Resolve configuration bugs, data races, registry lock contention, and typos - Fix slice pre-allocation bug in config loader to prevent registering empty devices. - Implement atomic file saving using temporary files to avoid file corruption. - Fix concurrency data race on sequenceNumber in IPv4 tracker using sync/atomic. - Reduce registry lock contention by saving devices with RLock. - Clone device property maps instead of sharing references to prevent map data races. - Correct minor "No suck tracker" typo.

+++ internal/trackers/ipv4/tracker_test.go
@@ -15,7 +15,7 @@ func TestNew(t *testing.T) {
 	tracker := tr.(*ipTracker)
 	assert.Equal(t, 5, tracker.pingPacketCount)
 	assert.Equal(t, 100*time.Millisecond, tracker.pingPacketDelay)
-	assert.Equal(t, 0, tracker.sequenceNumber)
+	assert.Equal(t, int32(0), tracker.sequenceNumber)
 
 	cfg["ping_packet_count"] = "1"
 	cfg["ping_packet_delay"] = "250ms"
